Rotate node log output
The console output of each log is split up in chunks of configurable
size, compressed, and when the size of the compressed output exceeds
a configurable limit, the oldest chunks are deleted.

+++ deployment/scripts/start-validator.sh
@@ -1,7 +1,24 @@
 #!/bin/bash
 
+# Obtain number of lines per log file.
+log_file_lines="$1"
+[ "${log_file_lines}" -gt 0 ] || exit
+shift
+
+# Make sure that the maximal log archive size (in bytes) has been properly specified.
+max_archive_size=$1
+[ "${max_archive_size}" -gt 0 ] || exit
+shift
+
 cd lotus || exit
 
+# Create log directory
+log_dir=~/spacenet-logs/validator-$(date +%Y-%m-%d-%H-%M-%S_%Z)
+mkdir -p "$log_dir"
+
+# Kill a potentially running validator.
 tmux kill-session -t mir-validator
 tmux new-session -d -s mir-validator
-tmux send-keys "./mir-validator run --nosync 2>&1" C-m
+
+# Start the Mir validator.
+tmux send-keys "./mir-validator run --nosync 2>&1 | ./rotate-logs.sh ${log_dir} ${log_file_lines} ${max_archive_size}" C-m
